updated the temperature measurements

+++ DSP2803x_examples_ccsv5/i2c_eeprom/User/User_Globals.c
@@ -197,7 +197,7 @@ volatile long testtrip = 0;
 volatile long timecounter = 0;
 volatile long timecounterseconds = 0;
 
-volatile float Charging_animation = 0;
+volatile Uint16 Charging_animation = 0;
 
 volatile Uint32 PreCharge_Timer = 0;
 volatile Uint16 Charger_Error_flag = 0;
