+++ DSP2803x_examples_ccsv5/i2c_eeprom/User/User_Globals.h
@@ -190,7 +190,7 @@ extern volatile long testtrip;
 extern volatile long timecounter;
 extern volatile long timecounterseconds;
 
-extern volatile float Charging_animation;
+extern volatile Uint16 Charging_animation;
 
 extern volatile Uint32 PreCharge_Timer;
 extern volatile Uint16 Charger_Error_flag;
